backend/requirements.txt
fastapi>=0.100.0
pydantic>=2.0.1
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.0,<1.5.0
alembic==1.13.1
psycopg2-binary>=2.9.0,<3.0.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-multipart>=0.0.5,<0.0.6
email-validator>=2.0.0
python-dotenv>=0.21.0
pydantic-settings>=2.0.0
cryptography>=41.0.0,<42.0.0
bcrypt==3.2.2

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.1
httpx>=0.19.0,<0.20.0

# Database Migrations
alembic==1.13.1 